Move mousetrackingannotation to services
